顶部ActionBar效果完成。

+++ app/src/main/java/com/blueinker/bluetrans/utils/ViewUtils.java
@@ -10,11 +10,11 @@ import android.util.TypedValue;
 public class ViewUtils {
 
 
-    public static float dp2px(Context context, float dpValue) {
-        return TypedValue.applyDimension(TypedValue.COMPLEX_UNIT_DIP, dpValue, context.getResources().getDisplayMetrics());
+    public static int dp2px(Context context, float dpValue) {
+        return (int)TypedValue.applyDimension(TypedValue.COMPLEX_UNIT_DIP, dpValue, context.getResources().getDisplayMetrics());
     }
 
-    public static float sp2px(Context context, float spValue) {
-        return TypedValue.applyDimension(TypedValue.COMPLEX_UNIT_SP, spValue, context.getResources().getDisplayMetrics());
+    public static int sp2px(Context context, float spValue) {
+        return (int)TypedValue.applyDimension(TypedValue.COMPLEX_UNIT_SP, spValue, context.getResources().getDisplayMetrics());
     }
 }
